feat: add image as string in props
